% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptcode.R
\name{testfunction}
\alias{testfunction}
\title{My Starter Code}
\usage{
testfunction(x)
}
\arguments{
\item{x}{is the numerical value you want}
}
\value{
returns the given x value multiplied by itself \code{\link{print}}
}
\description{
My Starter Code
}
\examples{
testfunction(8)
\dontrun{
testfunction(7)
}
}
